refactor: 把部分嵌套类抽取出来
