+++ memos/dev-log.md
@@ -2,6 +2,10 @@
 
 2024 年 6 月我才想起来写这玩意...
 
+## 2025.12.7
+
+
+
 ## 2024.7.30
 
 1. 考虑一种情况：插件在某次做增量备份前，用户热重载了插件配置，在里面新增了一个世界 `myworld`
